requirements.txt
autopep8==1.5.2
bandit==1.6.2
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
gitdb==4.0.5
GitPython==3.1.2
idna==2.9
pbr==5.4.5
pycodestyle==2.6.0
pycparser==2.20
python-telegram-bot==12.7
PyYAML==5.3.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
smmap==3.0.4
soupsieve==2.0
stevedore==1.32.0
tornado==6.0.4
urllib3==1.25.9
